+++ jenkins.sh
@@ -8,6 +8,8 @@ bundle install --path "${HOME}/bundles/${JOB_NAME}"
 bundle exec bin/govuk-lint-ruby lib bin/govuk-lint-ruby
 bundle exec bin/govuk-lint-ruby lib bin/govuk-lint-ruby --diff
 
+bundle exec rake test
+
 if [[ -n "$PUBLISH_GEM" ]]; then
   bundle exec rake publish_gem
 fi